Added ReservationService and moved there some methods from AccountService
