+++ src/main/java/ru/orion/library/services/ReservationService.java
@@ -1,0 +1,12 @@
+package ru.orion.library.services;
+
+import ru.orion.library.models.Account;
+import ru.orion.library.models.Book;
+
+import java.time.LocalDate;
+
+public interface ReservationService {
+    void reserve(Account account, Book book, LocalDate dateEnd);
+    void cancelReservation(Account account, Book book);
+    void extendReservation(Account account, Book book, LocalDate newDate);
+}
